dom/manifest/test/test_ManifestProcessor_warnings.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1086997
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1086997</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script src="common.js"></script>
  <script>
"use strict";
const options = {...data, checkConformance: true } ;
[
  {
    func: () => options.jsonText = JSON.stringify(1),
    warn: "Manifest should be an object.",
  },
  {
    func: () => options.jsonText = JSON.stringify("a string"),
    warn: "Manifest should be an object.",
  },
  {
    func: () => options.jsonText = JSON.stringify({
      scope: "https://www.mozilla.org",
    }),
    warn: "The scope URL must be same origin as document.",
  },
  {
    func: () => options.jsonText = JSON.stringify({
      scope: "foo",
      start_url: "bar",
    }),
    warn: "The start URL is outside the scope, so the scope is invalid.",
  },
  {
    func: () => options.jsonText = JSON.stringify({
      start_url: "https://www.mozilla.org",
    }),
    warn: "The start URL must be same origin as document.",
  },
  {
    func: () => options.jsonText = JSON.stringify({
      start_url: 42,
    }),
    warn: "Expected the manifest\u2019s start_url member to be a string.",
  },
  {
    func: () => options.jsonText = JSON.stringify({
      theme_color: "42",
    }),
    warn: "theme_color: 42 is not a valid CSS color.",
  },
  {
    func: () => options.jsonText = JSON.stringify({
      background_color: "42",
    }),
    warn: "background_color: 42 is not a valid CSS color.",
  },
  {
    func: () => options.jsonText = JSON.stringify({
      icons: [
        { "src": "http://exmaple.com", "sizes": "48x48"},
        { "src": "http://:Invalid", "sizes": "48x48"},
      ],
    }),
    warn: "icons item at index 1 is invalid. The src member is an invalid URL http://:Invalid",
  },
].forEach((test, index) => {
  test.func();
  const result = processor.process(options);
  const [message] = result.moz_validation;
  is(message.warn, test.warn, "Check warning.");
  options.manifestURL = manifestURL;
  options.docURL = docURL;
});

  </script>
</head>
